Move unchecked formatting to an unchecked module
