+++ components/datetime/src/lib.rs
@@ -105,15 +105,14 @@ pub mod provider;
 pub(crate) mod raw;
 pub mod scaffold;
 pub(crate) mod size_test_macro;
+pub mod unchecked;
 
 pub use error::{DateTimeFormatterLoadError, MismatchedCalendarError};
 
-pub use format::DateTimeInputUnchecked;
 pub use neo::DateTimeFormatter;
 pub use neo::DateTimeFormatterPreferences;
 pub use neo::FixedCalendarDateTimeFormatter;
 pub use neo::FormattedDateTime;
-pub use neo::FormattedDateTimeUnchecked;
 pub use neo::NoCalendarFormatter;
 
 /// Locale preferences used by this crate
